Refactor: split players into players and actors
The new module actors.py does most of the things
that players.py used to do. When creating a new
player (from a user_id), first an actor is created
(actor_id = player_id).
Handles, finances and chats all work based on
actors, not players.
Currently the only functionality of player is:
- Map the actor to the discord user. The actor makes
  sure there is a role that has all the correct
  permissions; the player takes the last step of
  giving that role to the user.
- Create the personal cmd_line channel for the user.

Motivation: we can now make Shops be another type of
actor. Then the shop will get financial records, a
handle, and a role for permissions. Then certain
players (or even handles, for in-game coop shops)
can be assigned as owners/workers at the shop, and
automatically gain access to its finances, orders
and chats.