many-to-many relation shops<->players
Refactoring in order to allow each shop
to have more than one player (employee)
and each player to be connected to more
than one shop.
The latter is not likely to be relevant,
but it comes with the territory of
having a mapping from player to shop,
which is necessary to avoid the
employees having to specify the name of
their shop for every shop command they
do.