Most shop commands require user to be employee
Most of the shop commands are no longer
gm-only, but instead check that the player is
indeed an employee of the shop in question.
In addition, the commands are made to work
if no shop name is specified, by taking the
first shop that the player is employed at
(if any). We don't expect anyone to be employed
at more than one shop, for the first LARP
at least.