Refactoring catalogue mapping for shops
When processing a reaction in a storefront
channel, we already have the mapping
channel_id -> shop. This means that the mapping
msg_id -> (shop, product) could actually be
reduced to just msg_id -> product, and moved
to the shop-specific catalogue file instead
of the shared shops.conf.

The reduction of the class CatalogueItemMapping
to just a product ID is not done in this
commit.