Check timestamp of previous order
When a new order comes in, we check the
prevous "active order" for that delivery ID.
If it is too old, we move the old order to
the "locked" category. If not, we combine
its contents with the new order.

TODO: currently, we always post the new
combined order as a new message in the order_flow
chat. We should either edit the message or
delete+repost.