Added groups
Commands to:
- Create a new group (with creator as first member)
- Add a member (based on handle)
- Clear all groups
- TODO: add command to remove member from group

Each group gets a channel under "private networks".
It works as a regular pseudonymous channel (based
on handle as usual), but only visible to the group.

TODO: Add functionality in player_config to put
players in starting groups automatically
TODO: perhaps an open command for inviting new
people to your group? Or let that be GM-only?
(probably best to let spontaneous groupings be
handled using multi-party chats -- or else we
would need a channel budget for people creating
new groups...)