Finances refactor + records log file
Refactor finances conf file to use a separate file
for each handle. This should prevent most
read-write collisions. Performing multiple
operations to the same handle at the same time is
still not thread-safe, however.

Also add an internal financial record to the same
file, meaning that there is a persistent financial
for each handle outside of the actor's "finance"
channel (which collects the records for all the
actor's handles). It is currently not used for
anything, but in the future I am planning a
"hacking" feature where another player might access
someone's financial record.