Add GM actor
The GM actor has shared handles for the GMs.
So far, any shared handles cannot be actively
used since they do not belong to the actor of
any player. They are only useful for chats
initiated by other players, which can then
be seen (and responded to) by all GMs.