Auto create categories on server.
All categories are created without any specific role permissions.
This functionality is just to make it easier to get going faster
instead of having to type each channel name manually.
Declare categories in the all_categories list (see common.py) in
the order you want the categories to appear on the server.